Factor plugin helper to dedicated module.
